+++ src/lib/ui/add-product-steps/BasicInfoStep.svelte
