+++ src/lib/ui/add-product-steps/DesktopStepNavigation.svelte
