+++ src/lib/ui/add-product-steps/ImagesStep.svelte
