+++ src/lib/ui/add-product-steps/IngredientsStep.svelte
