+++ src/lib/ui/add-product-steps/MobileStepHeader.svelte
