+++ src/lib/ui/add-product-steps/NavigationButtons.svelte
