+++ src/lib/ui/add-product-steps/NutritionStep.svelte
