+++ src/lib/ui/edit-product-steps/BasicInfoStep.svelte
