+++ src/lib/ui/edit-product-steps/ImagesStep.svelte
